+++ src/support/components/FreeSupportMessage.tsx
@@ -1,0 +1,10 @@
+import React, {FC} from 'react'
+
+export const FreeSupportMessage: FC = () => {
+  return (
+    <p>
+      Support is available only for paying customers. Please upgrade your
+      account or log into your paid account to submit a support case.
+    </p>
+  )
+}
